+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/authn/requests/ListGroupsRequest.java
@@ -1,0 +1,30 @@
+package cloud.pangeacyber.pangea.authn.requests;
+
+import cloud.pangeacyber.pangea.BaseRequest;
+import cloud.pangeacyber.pangea.authn.models.GroupsFilter;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonInclude.Include;
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import lombok.Builder;
+import lombok.Value;
+import lombok.experimental.SuperBuilder;
+import lombok.extern.jackson.Jacksonized;
+
+@Jacksonized
+@JsonIgnoreProperties(ignoreUnknown = true)
+@JsonInclude(Include.NON_NULL)
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+@SuperBuilder
+@Value
+public final class ListGroupsRequest extends BaseRequest {
+
+	GroupsFilter filter;
+	String last;
+	String order;
+	String orderBy;
+
+	@Builder.Default
+	Integer size = null;
+}