+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/authn/requests/UpdateGroupRequest.java
@@ -1,0 +1,30 @@
+package cloud.pangeacyber.pangea.authn.requests;
+
+import cloud.pangeacyber.pangea.BaseRequest;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonInclude.Include;
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import java.util.Map;
+import lombok.NonNull;
+import lombok.Value;
+import lombok.experimental.SuperBuilder;
+import lombok.extern.jackson.Jacksonized;
+
+@Jacksonized
+@JsonIgnoreProperties(ignoreUnknown = true)
+@JsonInclude(Include.NON_NULL)
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+@SuperBuilder
+@Value
+public final class UpdateGroupRequest extends BaseRequest {
+
+	@NonNull
+	String id;
+
+	String name;
+	String description;
+	String type;
+	Map<String, String> attributes;
+}