+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - AuthZ: `expiresAt` to tuples.
+- AuthN: groups.
 
 ## 5.0.0 - 2025-04-21